requirements/base.txt
Django == 4.2.11
django-environ==0.11.2
psycopg2==2.9.9
djangorestframework==3.15.1

requests
celery==5.3.6
django-celery-beat==2.6.0
pillow
python-magic
djangorestframework_simplejwt
cryptography
django-allauth
requests-oauthlib
python3-openid
pyjwt
python3-saml
setuptools
drf-yasg
djangorestframework-serializer-extensions
drf-writable-nested
